+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/property/PropertiesRegistry.java
@@ -46,6 +46,7 @@ public class PropertiesRegistry {
         register("fire", MPRFireProperty.class);
         register("freeze", MPRFreezeProperty.class);
         register("heal", MPRHealProperty.class);
+        register("function", MPRFunctionProperty.class);
         register("event", MPREventProperty.class);
         register("presets", MPRPresetsProperty.class);
         if (ModList.get().isLoaded("pehkui"))