+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/property/events/MPRKillEvent.java
@@ -4,7 +4,6 @@ import com.google.gson.*;
 import com.google.gson.annotations.JsonAdapter;
 import insane96mcp.mobspropertiesrandomness.data.json.condition.MPRCondition;
 import insane96mcp.mobspropertiesrandomness.data.json.property.MPRProperty;
-import net.minecraft.commands.CommandFunction;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.GsonHelper;
 import net.minecraft.world.damagesource.DamageSource;
@@ -19,8 +18,8 @@ import java.util.List;
 public class MPRKillEvent extends MPREvent {
 	public MPRHurtData hurtData;
 
-	public MPRKillEvent(MPRHurtData hurtData, ResourceLocation id, Target target, @Nullable CommandFunction.CacheableFunction function, @Nullable List<MPRProperty> applyProperties, List<MPRCondition> conditions) {
-		super(id, target, function, applyProperties, conditions);
+	public MPRKillEvent(MPRHurtData hurtData, ResourceLocation id, Target target, @Nullable List<MPRProperty> applyProperties, List<MPRCondition> conditions) {
+		super(id, target, applyProperties, conditions);
 		this.hurtData = hurtData;
 	}
 
@@ -47,11 +46,10 @@ public class MPRKillEvent extends MPREvent {
 			JsonObject jObject = json.getAsJsonObject();
 			String id = GsonHelper.getAsString(jObject, "id");
 			Target target = GsonHelper.getAsObject(jObject, "target", context, Target.class);
-			CommandFunction.CacheableFunction function = deserializeFunction(jObject);
 			List<MPRCondition> conditions = MPRCondition.deserializeConditions(jObject, context);
 			List<MPRProperty> properties = MPRProperty.deserializeList(jObject, "apply_properties", context);
 
-			return new MPRKillEvent(MPRHurtData.deserialize(jObject, context), ResourceLocation.parse(id), target, function, properties, conditions);
+			return new MPRKillEvent(MPRHurtData.deserialize(jObject, context), ResourceLocation.parse(id), target, properties, conditions);
 		}
 
 		@Override